--- a/statement.go
+++ b/statement.go
@@ -74,7 +74,7 @@ func (stmt *mysqlStmt) Exec(args []driver.Value) (driver.Result, error) {
 		// Columns
 		if metadataFollows && stmt.mc.extCapabilities&clientCacheMetadata != 0 {
 			// we can not skip column metadata because next stmt.Query() may use it.
-			if stmt.columns, err = mc.readColumns(resLen); err != nil {
+			if stmt.columns, err = mc.readColumns(resLen, stmt.columns); err != nil {
 				return nil, err
 			}
 		} else {
@@ -125,7 +125,7 @@ func (stmt *mysqlStmt) query(args []driver.Value) (*binaryRows, error) {
 	if resLen > 0 {
 		rows.mc = mc
 		if metadataFollows {
-			if rows.rs.columns, err = mc.readColumns(resLen); err != nil {
+			if rows.rs.columns, err = mc.readColumns(resLen, stmt.columns); err != nil {
 				return nil, err
 			}
 			stmt.columns = rows.rs.columns